project-loader: extracted libraries search into a separate file, use detectLogging(), add Java toolchain to build
